Extract storage interface
