+++ core/Storage/Filesystem.php
@@ -11,24 +11,51 @@
  * http://opensource.org/licenses/osl-3.0.php
  *
  * @category  Core
- * @package   Filesystem
+ * @package   Storage
  * @author    Daniel Kocherga <dan.kocherga@gmail.com>
  * @copyright 2013 Daniel Kocherga (dan.kocherga@gmail.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/dankocherga/MTool
  */
 
-namespace Core\Filesystem;
+namespace Core\Storage;
 
 /**
  * File-based filesystem
  *
  * @category Core
- * @package  Filesystem
+ * @package  Storage
  * @author   Daniel Kocherga <dan@oggettoweb.com>
  * @license  http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link     https://github.com/dankocherga/MTool
  */
-class Exception extends \Exception
+class Filesystem implements IStorage
 {
+    /**
+     * Create directory recursive 
+     * Separate directories with slash(/)
+     * 
+     * @param string $path Path
+     *
+     * @return void
+     */
+    public function mkdir($path)
+    {
+        $preparedPath = str_replace('/', DIRECTORY_SEPARATOR, $path);
+        if (!mkdir($preparedPath, 0755, true)) {
+            throw new Exception("Cannot create directory '{$preparedPath}'");
+        }
+    }
+
+    /**
+     * Write data to file
+     * 
+     * @param string $path    Path
+     * @param string $content Content
+     *
+     * @return void
+     */
+    public function write($path, $content)
+    {
+    }
 }
