+++ core/Storage/IStorage.php
@@ -11,25 +11,25 @@
  * http://opensource.org/licenses/osl-3.0.php
  *
  * @category  Core
- * @package   Filesystem
+ * @package   Storage
  * @author    Daniel Kocherga <dan.kocherga@gmail.com>
  * @copyright 2013 Daniel Kocherga (dan.kocherga@gmail.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/dankocherga/MTool
  */
 
-namespace Core;
+namespace Core\Storage;
 
 /**
  * Filesystem interface
  *
  * @category Core
- * @package  Filesystem
+ * @package  Storage
  * @author   Daniel Kocherga <dan@oggettoweb.com>
  * @license  http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link     https://github.com/dankocherga/MTool
  */
-interface IFilesystem
+interface IStorage
 {
     /**
      * Create directory recursive
